requirements.txt
streamlit
pandas
psycopg2-binary
requests
openpyxl
playwright
streamlit
playwright
pandas
beautifulsoup4
playwright
playwright
playwright-stealth
